+++ src/ms/symbol.js
@@ -56,6 +56,8 @@ function Symbol() {
   this.style.outlineWidth = 0; // Width of the outline.
   this.style.icon = true; // Should we display the icon?
   this.style.iconColor = "";
+  this.style.infoBackground = ""; // Color of square behind texts
+  this.style.infoBackgroundFrame = ""; // Color of the squares frame
   this.style.monoColor = ""; // Should the icon be monocromatic and if so what color
   this.style.civilianColor = true; // Should we use the Civilian Purple defined in 2525? (We set this to default because I like the color.
   this.style.colorMode = "Light"; // 2525C Allows you to use Dark; Medium or Light colors. The values you can set are "Dark";"Medium" or "Light"