+++ src/symbolfunctions/textfields.js
@@ -12,6 +12,15 @@ export default function textfields(ms) {
   var fontFamily = this.style.fontfamily;
   var fontSize = this.style.infoSize;
 
+  var infoBackground =
+    typeof this.style.infoBackground === "object"
+      ? this.style.infoBackground[this.metadata.affiliation]
+      : this.style.infoBackground;
+  var infoBackgroundFrame =
+    typeof this.style.infoBackground === "object"
+      ? this.style.infoBackground[this.metadata.affiliation]
+      : this.style.infoBackground;
+
   var gbbox = new ms.BBox();
   var spaceTextIcon = 20; //The distance between the Icon and the labels
 
@@ -632,6 +641,141 @@ export default function textfields(ms) {
       gbbox.y2 = Math.max(gbbox.y2, 100 + 2.7 * fontSize);
     }
 
+    // Background boxes behind text
+    if (this.style.infoBackground) {
+      var leftBox = { x1: 100, y1: 1000, y2: 0 };
+      var rightBox = { x1: 100, y1: 1000, y2: 0 };
+      if (gStrings.L1)
+        leftBox = {
+          x1: Math.min(leftBox.x1, bbox.x1 - strWidth(gStrings.L1)),
+          x2: bbox.x1 - spaceTextIcon / 2,
+          y1: Math.min(leftBox.y1, 100 - 2.5 * fontSize),
+          y2: Math.max(leftBox.y2, 100 - 1.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (gStrings.L2)
+        leftBox = {
+          x1: Math.min(leftBox.x1, bbox.x1 - strWidth(gStrings.L2)),
+          x2: bbox.x1 - spaceTextIcon / 2,
+          y1: Math.min(leftBox.y1, 100 - 1.5 * fontSize),
+          y2: Math.max(leftBox.y2, 100 - 0.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (gStrings.L3)
+        leftBox = {
+          x1: Math.min(leftBox.x1, bbox.x1 - strWidth(gStrings.L3)),
+          x2: bbox.x1 - spaceTextIcon / 2,
+          y1: Math.min(leftBox.y1, 100 - 0.5 * fontSize),
+          y2: Math.max(leftBox.y2, 100 + 0.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (gStrings.L4)
+        leftBox = {
+          x1: Math.min(leftBox.x1, bbox.x1 - strWidth(gStrings.L4)),
+          x2: bbox.x1 - spaceTextIcon / 2,
+          y1: Math.min(leftBox.y1, 100 + 0.5 * fontSize),
+          y2: Math.max(leftBox.y2, 100 + 1.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (gStrings.L5)
+        leftBox = {
+          x1: Math.min(leftBox.x1, bbox.x1 - strWidth(gStrings.L5)),
+          x2: bbox.x1 - spaceTextIcon / 2,
+          y1: Math.min(leftBox.y1, 100 + 1.5 * fontSize),
+          y2: Math.max(leftBox.y2, 100 + 2.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (leftBox.hasOwnProperty("x2")) {
+        gbbox.x1 -= fontSize / 2;
+        drawArray2.push({
+          type: "path",
+          d:
+            "M " +
+            (leftBox.x1 - fontSize / 2) +
+            "," +
+            (leftBox.y1 + fontSize / 2) +
+            " " +
+            leftBox.x1 +
+            "," +
+            leftBox.y1 +
+            " " +
+            leftBox.x2 +
+            "," +
+            leftBox.y1 +
+            " " +
+            leftBox.x2 +
+            "," +
+            leftBox.y2 +
+            " " +
+            (leftBox.x1 - fontSize / 2) +
+            "," +
+            leftBox.y2 +
+            "z",
+          fill: infoBackground,
+          stroke: infoBackgroundFrame || false
+        });
+      }
+      if (gStrings.R1)
+        rightBox = {
+          x1: bbox.x2 + spaceTextIcon / 2,
+          x2: Math.max(rightBox.x1, bbox.x2 + strWidth(gStrings.R1)),
+          y1: Math.min(rightBox.y1, 100 - 2.5 * fontSize),
+          y2: Math.max(rightBox.y2, 100 - 1.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (gStrings.R2)
+        rightBox = {
+          x1: bbox.x2 + spaceTextIcon / 2,
+          x2: Math.max(rightBox.x1, bbox.x2 + strWidth(gStrings.R2)),
+          y1: Math.min(rightBox.y1, 100 - 1.5 * fontSize),
+          y2: Math.max(rightBox.y2, 100 - 0.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (gStrings.R3)
+        rightBox = {
+          x1: bbox.x2 + spaceTextIcon / 2,
+          x2: Math.max(rightBox.x1, bbox.x2 + strWidth(gStrings.R3)),
+          y1: Math.min(rightBox.y1, 100 - 0.5 * fontSize),
+          y2: Math.max(rightBox.y2, 100 + 0.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (gStrings.R4)
+        rightBox = {
+          x1: bbox.x2 + spaceTextIcon / 2,
+          x2: Math.max(rightBox.x1, bbox.x2 + strWidth(gStrings.R4)),
+          y1: Math.min(rightBox.y1, 100 + 0.5 * fontSize),
+          y2: Math.max(rightBox.y2, 100 + 1.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (gStrings.R5)
+        rightBox = {
+          x1: bbox.x2 + spaceTextIcon / 2,
+          x2: Math.max(rightBox.x1, bbox.x2 + strWidth(gStrings.R5)),
+          y1: Math.min(rightBox.y1, 100 + 1.5 * fontSize),
+          y2: Math.max(rightBox.y2, 100 + 2.5 * fontSize + spaceTextIcon / 2)
+        };
+      if (rightBox.hasOwnProperty("x2")) {
+        gbbox.x2 += fontSize / 2;
+        drawArray2.push({
+          type: "path",
+          d:
+            "M " +
+            rightBox.x1 +
+            "," +
+            rightBox.y1 +
+            " " +
+            (rightBox.x2 + fontSize / 2) +
+            "," +
+            rightBox.y1 +
+            " " +
+            (rightBox.x2 + fontSize / 2) +
+            "," +
+            (rightBox.y2 - fontSize / 2) +
+            " " +
+            rightBox.x2 +
+            "," +
+            rightBox.y2 +
+            " " +
+            rightBox.x1 +
+            "," +
+            rightBox.y2 +
+            "z",
+          fill: infoBackground,
+          stroke: infoBackgroundFrame || false
+        });
+      }
+    }
     //geometries
     if (gStrings.L1)
       drawArray2.push({